Fix 32-bit build
When trying to build for 32-bit ARM (linux/arm), you get:

```
pkg/remote/client.go:120:35: cannot use 0xffffffff (untyped int constant 4294967295) as int value in argument to fmt.Errorf (overflows)
```

Pass the constant as unsigned 64-bit integer, which is what the
underlying snappy code is doing (it's casting the argument to an uint64
and comparing against 0xffffffff, which in Go means it's comparing
against uint64(0xffffffff)). Since this is just an error message, it
shouldn't matter, but let's be consistent.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>